+++ tests/test_via.py
@@ -11,6 +11,7 @@ import mock
 import unittest
 
 from flask_via import Via, RoutesImporter
+from flask_via.exceptions import ImproperlyConfigured
 
 
 class TestVia(unittest.TestCase):
@@ -21,7 +22,7 @@ class TestVia(unittest.TestCase):
     def test_init_app_raises_not_implemented(self):
         via = Via()
 
-        with self.assertRaises(NotImplementedError) as e:
+        with self.assertRaises(ImproperlyConfigured) as e:
             via.init_app(self.app)
 
         self.assertEqual(
@@ -34,6 +35,16 @@ class TestVia(unittest.TestCase):
 
         with self.assertRaises(ImportError):
             via.init_app(self.app)
+
+    @mock.patch('flask_via.RoutesImporter.include')
+    def test_via_routes_name_app_config(self, _include):
+        via = Via()
+        self.app.config['VIA_ROUTES_MODULE'] = 'foo.bar'
+        self.app.config['VIA_ROUTES_NAME'] = 'urls'
+
+        via.init_app(self.app)
+
+        _include.assert_called_once_with('foo.bar', 'urls')
 
     @mock.patch('flask_via.import_module')
     def test_init_app_raises_attribute_error(self, import_module):